Escape correctly in the face of dollar signs (and anything else)
As a comment suggested and the 'lol=>lol' test from the previous
commit shows, the previous version would produce the wrong result in
some cases, specifically when the input contained a double dollar sign
-- and this could be turned into an injection by a clever attacker.

It'd be tricky to get this exactly right with dollar-quoting -- from
the docs, it seems to be a mechanism aimed mainly at handwritten SQL
code, which perhaps explains that.  Fortunately another string constant
syntax can be used reliably in PostgreSQL; see code and comment.