Differentiate parsing own output and PostgreSQL's
When .load is used for parsing data received from PostgreSQL,
it will be given the actual hstore string, not a SQL string constant
describing it.  So don't try to strip off the encoding we
would have performed to produce a SQL string constant.

In the previous version, this could have corrupted data if
a key began with a double dollar sign and PostgreSQL chose
to return the field without enclosing the key in double quotes.
Nothing in the documentation guarantees that PostgreSQL will
use double quotes when optional, though fortunately it empirically
seems to do so in the version I have handy to test with (9.1).

For tests which parse our own output instead of PostgreSQL's,
use for_parameter=true when producing that output.