specs: collect data, dedup code
As a bonus now that the code is deduplicated so it's easy,
test both the for_parameter=true and (default) for_parameter=false
cases.

The test labelled with @teeparham's name turns out to
be essentially the same as what we call "should support
double quotes in strings".  Say it just once, and prefer the
description of what it does to who originally contributed it.