Add failing test case for injection with dollar signs
As a comment suggests, we can produce the wrong result in some
cases, specifically when the input contains a double dollar sign.

As this test demonstrates, the bug can be turned into an injection
if a user can influence one of the values in the input and is clever.
NB the failure is demonstrated only in the live PostgreSQL test;
it's when PostgreSQL parses the query that things really go wrong.

Fix in next commit.